+++ cfgov/cfgov/test.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 
 import importlib
+import re
 
 from django.apps import apps
 from django.test.runner import DiscoverRunner
@@ -28,3 +29,15 @@ class TestDataTestRunner(DiscoverRunner):
 
         module = importlib.import_module(migration)
         module.initial_data(apps, None)
+
+
+class HtmlMixin(object):
+    def assertHtmlRegexpMatches(self, s, r):
+        s_no_right_spaces = re.sub('>\s*', '>', s)
+        s_no_left_spaces = re.sub('\s*<', '<', s_no_right_spaces)
+        s_no_extra_spaces = re.sub('\s+', ' ', s_no_left_spaces)
+
+        self.assertIsNotNone(
+            re.search(r, s_no_extra_spaces.strip(), flags=re.DOTALL),
+            '{} did not match {}'.format(s_no_extra_spaces, r)
+        )
