+++ cfgov/jinja2/v1/about-us/careers/_macro-table-current-openings.html
@@ -29,7 +29,7 @@
                 POSTING CLOSES
             </th>
             <th>
-                LOCATION
+                REGION
             </th>
         </tr>
     </thead>
@@ -42,8 +42,8 @@
             <td data-label="Posting Closes">
                 {{ time.render(career.close_date, {'date':true}) }}
             </td>
-            <td data-label="Location">
-                {{ career.location_names|join(', ') }}
+            <td data-label="Region">
+                {{ career.regions.order_by('region__region_long') | join(', ') }}
             </td>
         </tr>
     {% endfor %}
