+++ cfgov/jobmanager/models/__init__.py
@@ -1,3 +1,4 @@
+from .blocks import *
 from .django import *
 from .pages import *
 from .panels import *
