+++ cfgov/jobmanager/models/blocks.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from django.template.loader import render_to_string
 from django.utils import timezone
 from v1.atomic_elements.organisms import ModelList, ModelTable
